Moved classes responsible for handling command line arguments to CommandLineArguments folder
