+++ Tests/CommandLineParserTests.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.IO;
+using System.Linq;
+using Xunit;
+using FluentAssertions;
+using YoCode;
+
+namespace YoCode_XUnit
+{
+    public class CommandLineParserTests
+    {
+
+    }
+}
