+++ YoCode/CommandLineArguments/ArgErrorType.cs
@@ -1,0 +1,10 @@
+﻿namespace YoCode
+{
+    public enum ArgErrorType
+    {
+        NoArguments,
+        WrongOriginalDirectory,
+        WrongModifiedDirectory,
+        WrongCommand
+    }
+}
