+++ YoCode/CommandLineArguments/SplitArg.cs
@@ -1,0 +1,8 @@
+﻿namespace YoCode
+{
+    public struct SplitArg
+    {
+        public string command;
+        public string data;
+    }
+}
